backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic_settings
python-dotenv
requests
pydantic
aiohttp